Created base for grpah component
