+++ .idea/misc.xml
@@ -4,6 +4,12 @@
     <option name="filePathToZoomLevelMap">
       <map>
         <entry key="../../../../../layout/compose-model-1644847396139.xml" value="0.5" />
+        <entry key="../../../../../layout/compose-model-1644865279641.xml" value="0.34814814814814815" />
+        <entry key="../../../../../layout/compose-model-1644915581771.xml" value="0.3398148148148148" />
+        <entry key="../../../../../layout/compose-model-1644918686739.xml" value="0.32559121621621623" />
+        <entry key="../../../../../layout/compose-model-1644919494445.xml" value="0.28055555555555556" />
+        <entry key="../../../../../layout/compose-model-1644919781084.xml" value="0.1" />
+        <entry key="app/src/main/res/layout/graph.xml" value="0.37604166666666666" />
       </map>
     </option>
   </component>
