+++ app/src/main/java/fi/sportionbois/sportion/MainActivity.kt
@@ -1,28 +1,21 @@
 package fi.sportionbois.sportion
 
 import android.os.Bundle
+import android.util.Log
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.compose.foundation.layout.*
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.material.*
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.*
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.sp
 import androidx.navigation.NavController
-import androidx.navigation.NavHostController
-import androidx.navigation.compose.NavHost
-import androidx.navigation.compose.composable
-import androidx.navigation.compose.currentBackStackEntryAsState
 import androidx.navigation.compose.rememberNavController
-import fi.sportionbois.sportion.navigation.BottomNavItem
+import fi.sportionbois.sportion.navigation.BottomNavigationBar
+import fi.sportionbois.sportion.navigation.NavigationGraph
+import fi.sportionbois.sportion.navigation.TopBar
 import fi.sportionbois.sportion.ui.theme.SportionTheme
 
 class MainActivity : ComponentActivity() {
@@ -41,23 +34,32 @@ class MainActivity : ComponentActivity() {
 
 @Composable
 fun Home() {
-    LazyColumn(modifier = Modifier .fillMaxHeight() .fillMaxWidth(),
-    horizontalAlignment = Alignment.CenterHorizontally,
-    verticalArrangement = Arrangement.SpaceEvenly){
-        items(2){
-            HomeDataCard(data = Message("Ville Valo", "Laulanta", "REPEMAX",
-                "1.4s", "50kg", "3"))
-            HomeDataCard(data = Message("Ville Valo2", "Laulanta", "REPEMAX",
-                "1.4s", "50kg", "3"))
+    CenteredColumnMaxWidthAndHeight {
+        LazyColumn() {
+            items(2) {
+                HomeDataCard(
+                    data = Message(
+                        "Ville Valo", "Laulanta", "REPEMAX",
+                        "1.4s", "50kg", "3"
+                    )
+                )
+                HomeDataCard(
+                    data = Message(
+                        "Ville Valo2", "Laulanta", "REPEMAX",
+                        "1.4s", "50kg", "3"
+                    )
+                )
+            }
         }
     }
-
 }
 
 @Composable
-fun StartTracking(name: String) {
-    ButtonCHViolet(text = "START") {
-        
+fun StartTracking(navController: NavController) {
+    Column(modifier = Modifier, horizontalAlignment = Alignment.CenterHorizontally) {
+        ButtonCHViolet(text = "START", onClick = {
+            navController.navigate("LiftDetails")
+        })
     }
 }
 
@@ -71,15 +73,9 @@ fun MainScreen() {
     val navController = rememberNavController()
     Scaffold(
         topBar = { TopBar() },
-        bottomBar = { BottomNavigation(navController) }
+        bottomBar = { BottomNavigationBar(navController) }
     ) {
-        Column(
-            modifier = Modifier
-                .fillMaxWidth()
-                .fillMaxHeight(),
-            horizontalAlignment = Alignment.CenterHorizontally,
-            verticalArrangement = Arrangement.Center,
-        ) {
+        CenteredColumnMaxWidthAndHeight {
             NavigationGraph(navController = navController)
         }
     }
@@ -91,86 +87,5 @@ fun MainScreen() {
 fun DefaultPreview() {
     SportionTheme {
         Home()
-    }
-}
-
-@Composable
-fun TopBar() {
-    TopAppBar(
-        backgroundColor = MaterialTheme.colors.background,
-        contentColor = MaterialTheme.colors.onBackground,
-        title = { Text(stringResource(id = R.string.app_name)) },
-        navigationIcon = {
-            IconButton(onClick = { /* doSomething() */ }) {
-                Icon(Icons.Filled.Menu, contentDescription = null)
-            }
-        },
-        actions = {
-            IconButton(onClick = { /* doSomething() */ }) {
-                Icon(Icons.Filled.Search, contentDescription = "Localized description")
-            }
-            IconButton(onClick = { /* doSomething() */ }) {
-                Icon(
-                    Icons.Filled.MoreVert,
-                    contentDescription = "Localized description"
-                )
-            }
-        }
-    )
-}
-
-@Composable
-fun NavigationGraph(navController: NavHostController) {
-    NavHost(navController = navController, startDestination = BottomNavItem.Home.screen_route) {
-        composable(BottomNavItem.Home.screen_route) {
-            Home()
-        }
-        composable(BottomNavItem.StartTracking.screen_route) {
-            StartTracking(name = "Start tracking")
-        }
-        composable(BottomNavItem.Settings.screen_route) {
-            Settings(name = "Settings")
-        }
-    }
-}
-
-@Composable
-fun BottomNavigation(navController: NavController) {
-    val items = listOf(
-        BottomNavItem.Home,
-        BottomNavItem.StartTracking,
-        BottomNavItem.Settings,
-    )
-    BottomNavigation(
-        backgroundColor = MaterialTheme.colors.background
-    ) {
-        val navBackStackEntry by navController.currentBackStackEntryAsState()
-        val currentRoute = navBackStackEntry?.destination?.route
-        items.forEach { item ->
-            BottomNavigationItem(
-                icon = { Icon(imageVector = item.icon, contentDescription = item.title) },
-                label = {
-                    Text(
-                        text = item.title,
-                        fontSize = 9.sp
-                    )
-                },
-                selectedContentColor = MaterialTheme.colors.onBackground,
-                unselectedContentColor = MaterialTheme.colors.onBackground.copy(0.4f),
-                alwaysShowLabel = true,
-                selected = currentRoute == item.screen_route,
-                onClick = {
-                    navController.navigate(item.screen_route) {
-                        navController.graph.startDestinationRoute?.let { screen_route ->
-                            popUpTo(screen_route) {
-                                saveState = true
-                            }
-                        }
-                        launchSingleTop = true
-                        restoreState = true
-                    }
-                }
-            )
-        }
     }
 }
