+++ app/src/main/java/fi/sportionbois/sportion/navigation/TopBar.kt
@@ -1,0 +1,35 @@
+package fi.sportionbois.sportion.navigation
+
+import androidx.compose.material.*
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.Menu
+import androidx.compose.material.icons.filled.MoreVert
+import androidx.compose.material.icons.filled.Search
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.res.stringResource
+import fi.sportionbois.sportion.R
+
+@Composable
+fun TopBar() {
+    TopAppBar(
+        backgroundColor = MaterialTheme.colors.background,
+        contentColor = MaterialTheme.colors.onBackground,
+        title = { Text(stringResource(id = R.string.app_name)) },
+        navigationIcon = {
+            IconButton(onClick = { /* doSomething() */ }) {
+                Icon(Icons.Filled.Menu, contentDescription = null)
+            }
+        },
+        actions = {
+            IconButton(onClick = { /* doSomething() */ }) {
+                Icon(Icons.Filled.Search, contentDescription = "Localized description")
+            }
+            IconButton(onClick = { /* doSomething() */ }) {
+                Icon(
+                    Icons.Filled.MoreVert,
+                    contentDescription = "Localized description"
+                )
+            }
+        }
+    )
+}
