+++ app/src/main/res/layout/graph.xml
@@ -1,0 +1,10 @@
+<?xml version="1.0" encoding="utf-8"?>
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent">
+    <com.github.mikephil.charting.charts.LineChart
+        android:layout_width ="match_parent"
+        android:layout_height ="match_parent"
+        android:id="@+id/graph" />
+
+</LinearLayout>
